+++ firebase-crashlytics/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Unreleased
+* [feature] Added an overload for `recordException` that allows logging additional custom
+  keys to the non fatal event [#3551]
+
+# 19.3.0
 * [fixed] Fixed inefficiency in the Kotlin `FirebaseCrashlytics.setCustomKeys` extension.
 * [fixed] Execute failure listener outside the main thread  [#6535]
+
+
+## Kotlin
+The Kotlin extensions library transitively includes the updated
+`firebase-crashlytics` library. The Kotlin extensions library has no additional
+updates.
 
 # 19.2.1
 * [changed] Updated protobuf dependency to `3.25.5` to fix